enhance(client): ハイライトをみつけるに統合
