+++ CHANGELOG.md
@@ -11,6 +11,11 @@ You should also include the user name that made the change.
 
 ## 12.x.x (unreleased)
 
+### Changes
+- ハイライトがみつけるに統合されました
+- カスタム絵文字ページはインスタンス情報ページに統合されました
+- 連合ページはインスタンス情報ページに統合されました
+
 ### Improvements
 - Server: Allow GET method for some endpoints @syuilo
 - Server: Add rate limit to i/notifications @tamaina
