+++ packages/client/src/router.ts
@@ -62,9 +62,6 @@ export const routes = [{
 	path: '/about-misskey',
 	component: page(() => import('./pages/about-misskey.vue')),
 }, {
-	path: '/featured',
-	component: page(() => import('./pages/featured.vue')),
-}, {
 	path: '/theme-editor',
 	component: page(() => import('./pages/theme-editor.vue')),
 	loginRequired: true,
